Move gate application to circuit class
