+++ strawberryfields/backends/fockbackend/circuit.py
@@ -23,10 +23,12 @@ from itertools import product
 import numpy as np
 from numpy import sqrt, pi
 from scipy.special import factorial as bang
+from numba import jit
 from numba.typed import List
 
 from . import ops
 
+def_type = np.complex128
 indices = string.ascii_lowercase
 MAX_MODES = len(indices) - 3
 
@@ -36,7 +38,7 @@ class Circuit():
     in the fock basis.
     """
 
-    def __init__(self, num, trunc, pure=True, do_checks=False, mode='blas'):
+    def __init__(self, num, trunc, pure=True, do_checks=False):
         r"""Class initializer.
 
         Args:
@@ -58,25 +60,7 @@ class Circuit():
         self._num_modes = num
         self._hbar = 2
         self._checks = do_checks
-        self._mode = mode
         self.reset(pure=pure, cutoff_dim=trunc)
-
-    def _apply_gate(self, mat, modes):
-        """Master gate application function. Selects between implementations based
-        on the `_mode` class parameter.
-
-        Args:
-            mat (array): The matrix to apply
-            modes (list<non-negative int>): The modes to apply `mat` to
-        """
-
-        args = [mat, self._state, self._pure, modes, self._num_modes, self._trunc]
-        if self._mode == 'blas':
-            self._state = ops.apply_gate_BLAS(*args)
-        elif self._mode == 'einsum':
-            self._state = ops.apply_gate_einsum(*args)
-        else:
-            raise NotImplementedError
 
     def _apply_channel(self, kraus_ops, modes):
         """Master channel application function. Applies a channel represented by
@@ -96,15 +80,9 @@ class Circuit():
 
         if len(kraus_ops) == 0:
             self._state = np.zeros([self._trunc for i in range(self._num_modes*2)], dtype=ops.def_type)
-        elif self._mode == 'blas':
-            states = [ops.apply_gate_einsum(k, np.copy(self._state), False, modes, self._num_modes, self._trunc)\
-                                    for k in kraus_ops]
+        else:
+            states = [self.apply_gate_BLAS(k, modes, pure=False) for k in kraus_ops]
             self._state = sum(states)
-        elif self._mode == 'einsum':
-            states = [ops.apply_gate_einsum(k, self._state, False, modes, self._num_modes, self._trunc)\
-                                    for k in kraus_ops]
-            self._state = sum(states)
-
 
     def reset(self, pure=None, cutoff_dim=None, num_subsystems=None):
         """Resets the simulation state.
@@ -134,6 +112,241 @@ class Circuit():
             self._state = ops.vacuumState(self._num_modes, self._trunc)
         else:
             self._state = ops.vacuumStateMixed(self._num_modes, self._trunc)
+
+    def apply_gate_BLAS(self, mat, modes, **kwargs):
+        """
+        Gate application based on custom indexing and matrix multiplication.
+        Assumes the input matrix has shape (out1, in1, ...).
+
+        This implementation uses indexing and BLAS. As per stack overflow,
+        einsum doesn't actually use BLAS but rather a c implementation. In theory
+        if reshaping is efficient this should be faster.
+        """
+        # pylint: disable=too-many-branches
+
+        state = kwargs.get("state", self._state)
+        pure = kwargs.get("pure", self._pure)
+        n = kwargs.get("n", self._num_modes)
+
+        size = len(modes)
+        dim = self._trunc**size
+        stshape = [self._trunc for i in range(size)]
+
+        # Apply the following matrix transposition:
+        # |m1><m1| |m2><m2| ... |mn><mn| -> |m1>|m2>...|mn><m1|<m2|...<mn|
+        transpose_list = [2*i for i in range(size)] + [2*i + 1 for i in range(size)]
+        matview = np.transpose(mat, transpose_list).reshape((dim, dim))
+
+        # checks if matview is diagonal and, if so, use faster contractions
+        diag = np.all(matview == np.diag(np.diagonal(matview)))
+
+        if pure:
+            if n == 1:
+                if diag:
+                    mat_diag = mat.diagonal()
+                    return np.multiply(mat_diag, state)
+                return np.dot(mat, state)
+
+            # Transpose the state into the following form:
+            # |psi> |mode[0]> |mode[1]> ... |mode[n]>
+            transpose_list = [i for i in range(n) if not i in modes] + modes
+            view = np.transpose(state, transpose_list)
+
+            # Apply matrix to each substate
+            ret = np.zeros([self._trunc for i in range(n)], dtype=def_type)
+            for i in product(*([range(self._trunc) for j in range(n - size)])):
+                if diag:
+                    mat_diag = matview.diagonal()
+                    ret[i] = np.multiply(mat_diag, view[i].ravel()).reshape(stshape)
+                else:
+                    ret[i] = np.dot(matview, view[i].ravel()).reshape(stshape)
+
+            # "untranspose" the return matrix ret
+            untranspose_list = [0] * len(transpose_list)
+            for i in range(len(transpose_list)): # pylint: disable=consider-using-enumerate
+                untranspose_list[transpose_list[i]] = i
+
+            return np.transpose(ret, untranspose_list)
+
+        # otherwise, the state is mixed
+        if n == 1:
+            if diag:
+                mat_diag = mat.diagonal().reshape(-1, 1)
+                return np.multiply(mat_diag, np.multiply(state, mat_diag.conj().T))
+            return np.dot(mat, np.dot(state, mat.conj().T))
+
+        # Transpose the state into the following form:
+        # |psi><psi||mode[0]>|mode[1]>...|mode[n]><mode[0]|<mode[1]|...<mode[n]|
+        transpose_list = [i for i in range(n*2) if not i//2 in modes]
+        transpose_list = transpose_list + [2*i for i in modes] + [2*i + 1 for i in modes]
+        view = np.transpose(state, transpose_list)
+
+        # Apply matrix to each substate
+        ret = np.zeros([self._trunc for i in range(n*2)], dtype=def_type)
+        for i in product(*([range(self._trunc) for j in range((n - size)*2)])):
+            if diag:
+                mat_diag = matview.diagonal().reshape(-1, 1)
+                ret[i] = np.multiply(mat_diag, np.multiply(view[i].reshape((dim, dim)), mat_diag.conj().T)).reshape(stshape + stshape)
+            else:
+                ret[i] = np.dot(matview, np.dot(view[i].reshape((dim, dim)), matview.conj().T)).reshape(stshape + stshape)
+
+        # "untranspose" the return matrix ret
+        untranspose_list = [0] * len(transpose_list)
+        for i in range(len(transpose_list)): # pylint: disable=consider-using-enumerate
+            untranspose_list[transpose_list[i]] = i
+
+        return np.transpose(ret, untranspose_list)
+
+    def apply_twomode_gate(self, mat, modes, gate="BSgate"):
+        """Applies a two-mode gate to a state.
+
+        Applies the specified two-mode gate to the state using custom tensor contractions and
+        the Numba compiler for faster application. Currently, only the beamsplitter and the
+        two-mode squeeze gate are supported.
+
+        Args:
+            mat (array[complex]): The numeric operator to be applied to the state, of shape `[trunc]*(2*n)`
+            modes (list[int]): the list of modes to which the operator is applied on
+            gate (str): The gate that is being applied. This argument determines the selection rules that
+                are used. Options are ``"BSgate"`` and ``"S2gate"``.
+
+        Keyword args:
+            state (array[complex]): The state that the operator is applied to
+            pure (bool): whether the state is pure or mixed
+            n (int): The total number of modes
+
+        Returns:
+            array[complex]: state after application of the two-mode operation
+        """
+        if self._pure:
+            t1 = modes[0]
+            t2 = modes[1]
+
+            # moves the modes on which the gate should be applied to the
+            # front indices of the state, for mode 1 and 2 respectively
+            # e.g. from [i1, i2, i3] --> [i2, i3, i1] if the gate is applied to modes 2 and 3
+            switch_list_1 = np.arange(self._num_modes)
+            switch_list_2 = np.arange(self._num_modes)
+            switch_list_1[[0, t1]] = switch_list_1[[t1, 0]]
+            switch_list_2[[1, t2]] = switch_list_2[[t2, 1]]
+
+            self._state = self._state.transpose(switch_list_1)
+            self._state = self._state.transpose(switch_list_2)
+
+            if gate == "BSgate":
+                self._state = self._apply_BS(mat, self._state, self._trunc)
+            elif gate == "S2gate":
+                self._state = self._apply_S2(mat, self._state, self._trunc)
+            else:
+                raise NotImplementedError("Currently, selection rules are only implemented for the BSgate "
+                                          "and the S2gate. The {} gate is not supported".format(gate))
+
+            self._state = self._state.transpose(switch_list_2)
+            ret = self._state.transpose(switch_list_1)
+        else:
+            t1 = 2 * modes[0]
+            t2 = 2 * modes[1]
+
+            # moves the modes on which the gate should be applied to the
+            # front indices of the state, for mode 1 and 2 respectively
+            # e.g. from [i1, j1, i2, j2, i3, j3] --> [i2, j2, i3, j3, i1, j1]
+            # if the gate is applied to modes 2 and 3
+            switch_list_1 = np.arange(2 * self._num_modes)
+            switch_list_2 = np.arange(2 * self._num_modes)
+            switch_list_1[[0, 1, t1, t1+1]] = switch_list_1[[t1, t1+1, 0, 1]]
+            switch_list_2[[0, 1, t2, t2+1]] = switch_list_2[[t2, t2+1, 0, 1]]
+
+            # puts the modes on which the gate should be applied together
+            # e.g. from [i2, j2, i3, j3, i1, j1] --> [i2, i3, j2, j3, i1, j1]
+            # if the gate is applied to modes 2 and 3 (which are already switched to the first two modes)
+            transpose_list = np.arange(2 * self._num_modes)
+            transpose_list[[t1+1, t2]] = transpose_list[[t2, t1+1]]
+
+            self._state = self._state.transpose(transpose_list)
+            self._state = self._state.transpose(switch_list_1)
+
+            if gate == "BSgate":
+                self._state = self._apply_BS(mat, self._state, self._trunc)
+                self._state = self._state.transpose(switch_list_1)
+
+                self._state = self._state.transpose(switch_list_2)
+                self._state = self._apply_BS(mat.conj(), self._state, self._trunc)
+
+            elif gate == "S2gate":
+                self._state = self._apply_S2(mat, self._state, self._trunc)
+                self._state = self._state.transpose(switch_list_1)
+
+                self._state = self._state.transpose(switch_list_2)
+                self._state = self._apply_S2(mat.conj(), self._state, self._trunc)
+
+            else:
+                raise NotImplementedError
+
+            self._state = self._state.transpose(switch_list_2)
+            ret = self._state.transpose(transpose_list)
+
+        return ret
+
+    # ignored in covtest (doesn't work well with the jit decorator)
+    @staticmethod
+    @jit(nopython=True)
+    def _apply_BS(mat, state, trunc):  # pragma: no cover
+        r"""Applies the BS gate to the first bra in state.
+
+        The beamsplitter matrix elements :math:`B_{ij}^{kl}` satisfy the selection
+        rules
+
+        .. math::
+            B_{ij}^{kl} \propto \delta_{i+j}^{k+l}
+
+        This implies that one can contract pure states (without loss of generality
+        we assume only two modes) using one less loop as follows:
+
+        .. math::
+            c'_{i, j} = \sum\limits_{k = \text{max}(1+i+j-n, 0)}
+                    ^{\text{min}(i+j, n-1) + 1}
+            B_{ik}^{j, i+j-k} c_{k, i+j-k}
+
+        where :math:`c_{kl}` is the tensor representing the two mode state and
+        :math:`n` is the cutoff in Fock space. Similarly for a mixed state it is
+        possible to do the update of the state using two less summations.
+        """
+        ret = np.zeros_like(state, dtype=np.complex128)
+        for i in range(trunc):
+            for j in range(trunc):
+                for k in range(max(1+i+j-trunc, 0), min(i+j, trunc-1) + 1):
+                    ret[i, j] += mat[i, k, j, i+j-k] * state[k, i+j-k]
+        return ret
+
+    # ignored in covtest (doesn't work well with the jit decorator)
+    @staticmethod
+    @jit(nopython=True)
+    def _apply_S2(mat, state, trunc):  # pragma: no cover
+        r"""Applies the S2 gate to the first bra in state.
+
+        The two-mode squeeze matrix elements :math:`T_{ij}^{kl}` satisfy the selection
+        rules
+
+        .. math::
+            T_{ij}^{kl} \propto \delta_{i-j}^{k-l}
+
+        This implies that one can contract pure states (without loss of generality
+        we assume only two modes) using one less loop as follows:
+
+        .. math::
+            c'_{i, k} = \sum\limits_{k = \text{max}(i-j, 0)}^{\text{min}(i-j, 0) + n}
+            T_{ij}^{k, k+j-i} c_{j, k+j-i}
+
+        where :math:`c_{jl}` is the tensor representing the two mode state and
+        :math:`n` is the cutoff in Fock space. Similarly for a mixed state it is
+        possible to do the update of the state using two less summations.
+        """
+        ret = np.zeros_like(state, dtype=np.complex128)
+        for i in range(trunc):
+            for j in range(trunc):
+                for k in range(max(i-j, 0), trunc + min(i-j, 0)):
+                    ret[i, k] += mat[i, j, k, k + j-i] * state[j, k + j-i]
+        return ret
 
     def norm(self):
         """returns the norm of the state"""
@@ -303,13 +516,15 @@ class Circuit():
         """
         Applies a phase shifter.
         """
-        self._apply_gate(ops.phase(theta, self._trunc), [mode])
+        mat = ops.phase(theta, self._trunc)
+        self._state = self.apply_gate_BLAS(mat, [mode])
 
     def displacement(self, alpha, mode):
         """
         Applies a displacement gate.
         """
-        self._apply_gate(ops.displacement(alpha, self._trunc), [mode])
+        mat = ops.displacement(alpha, self._trunc)
+        self._state = self.apply_gate_BLAS(mat, [mode])
 
     def beamsplitter(self, t, r, phi, mode1, mode2):
         """
@@ -321,46 +536,47 @@ class Circuit():
         modes.append(mode1)
         modes.append(mode2)
 
-        args = [mat, self._state, self._pure, modes, self._num_modes, self._trunc]
-        self._state = ops.apply_twomode_gate(*args, gate="BSgate")
+        self._state = self.apply_twomode_gate(mat, modes, gate="BSgate")
 
     def squeeze(self, r, theta, mode):
         """
         Applies a squeezing gate.
         """
-        self._apply_gate(ops.squeezing(r, theta, self._trunc), [mode])
+        mat = ops.squeezing(r, theta, self._trunc)
+        self._state = self.apply_gate_BLAS(mat, [mode])
 
     def two_mode_squeeze(self, r, theta, mode1, mode2):
         """
         Applies a two-mode squeezing gate.
         """
-        # self._apply_gate(ops.two_mode_squeezing(r, theta, self._trunc), [mode1, mode2])
         mat = ops.two_mode_squeezing(r, theta, self._trunc)
 
         modes = List()
         modes.append(mode1)
         modes.append(mode2)
 
-        args = [mat, self._state, self._pure, modes, self._num_modes, self._trunc]
-        self._state = ops.apply_twomode_gate(*args, gate="S2gate")
+        self._state = self.apply_twomode_gate(mat, modes, gate="S2gate")
 
     def kerr_interaction(self, kappa, mode):
         """
         Applies a Kerr interaction gate.
         """
-        self._apply_gate(ops.kerr(kappa, self._trunc), [mode])
+        mat = ops.kerr(kappa, self._trunc)
+        self._state = self.apply_gate_BLAS(mat, [mode])
 
     def cross_kerr_interaction(self, kappa, mode1, mode2):
         """
         Applies a cross-Kerr interaction gate.
         """
-        self._apply_gate(ops.cross_kerr(kappa, self._trunc), [mode1, mode2])
+        mat = ops.cross_kerr(kappa, self._trunc)
+        self._state = self.apply_gate_BLAS(mat, [mode1, mode2])
 
     def cubic_phase_shift(self, gamma, mode):
         """
         Applies a cubic phase shift gate.
         """
-        self._apply_gate(ops.cubicPhase(gamma, self._hbar, self._trunc), [mode])
+        mat = ops.cubicPhase(gamma, self._hbar, self._trunc)
+        self._state = self.apply_gate_BLAS(mat, [mode])
 
     def is_vacuum(self, tol):
         """
@@ -492,11 +708,7 @@ class Circuit():
             reduced = ops.partial_trace(state, self._num_modes, unmeasured)
 
             # Rotate to measurement basis
-            args = [ops.phase(-phi, self._trunc), reduced, False, [0], 1, self._trunc]
-            if self._mode == 'blas':
-                reduced = ops.apply_gate_BLAS(*args)
-            elif self._mode == 'einsum':
-                reduced = ops.apply_gate_einsum(*args)
+            reduced = self.apply_gate_BLAS(ops.phase(-phi, self._trunc), [0], state=reduced, pure=False, n=1)
 
             # Create pdf. Same as tf implementation, but using
             # the recursive relation H_0(x) = 1, H_1(x) = 2x, H_{n+1}(x) = 2xH_n(x) - 2nH_{n-1}(x)
@@ -530,15 +742,12 @@ class Circuit():
         alpha = homodyne_sample * sqrt(m_omega_over_hbar / 2)
 
         composed = np.dot(ops.phase(phi, self._trunc), ops.displacement(alpha, self._trunc))
-        args = [composed, inf_squeezed_vac, True, [0], 1, self._trunc]
-        if self._mode == 'blas':
-            eigenstate = ops.apply_gate_BLAS(*args)
-        elif self._mode == 'einsum':
-            eigenstate = ops.apply_gate_einsum(*args)
+        eigenstate = self.apply_gate_BLAS(composed, [0], state=inf_squeezed_vac, pure=True, n=1)
 
         vac_state = np.array([1.0 + 0.0j if i == 0 else 0.0 + 0.0j for i in range(self._trunc)], dtype=ops.def_type)
         projector = np.outer(vac_state, eigenstate.conj())
-        self._apply_gate(projector, [mode])
+
+        self._state = self.apply_gate_BLAS(projector, [mode])
 
         # Normalize
         self._state = self._state / self.norm()
