+++ strawberryfields/backends/fockbackend/ops.py
@@ -27,8 +27,6 @@ from numpy.polynomial.hermite import hermval as H
 from scipy.special import factorial as fac
 from scipy.linalg import expm as matrixExp
 
-from numba import jit
-
 from thewalrus.fock_gradients import Dgate, Sgate, S2gate, BSgate
 
 
@@ -205,287 +203,6 @@ def project_reset(modes, x, state, pure, n, trunc):
         ret = np.zeros([trunc for i in range(n*2)], dtype=def_type)
         ret[intersperse(outSlice)] = state[intersperse(inSlice)]
 
-    return ret
-
-
-# ============================================
-#
-# Matrix multiplication
-#
-# ============================================
-
-def apply_gate_BLAS(mat, state, pure, modes, n, trunc):
-    """
-    Gate application based on custom indexing and matrix multiplication.
-    Assumes the input matrix has shape (out1, in1, ...).
-
-    This implementation uses indexing and BLAS. As per stack overflow,
-    einsum doesn't actually use BLAS but rather a c implementation. In theory
-    if reshaping is efficient this should be faster.
-    """
-    # pylint: disable=too-many-branches
-
-    size = len(modes)
-    dim = trunc**size
-    stshape = [trunc for i in range(size)]
-
-    # Apply the following matrix transposition:
-    # |m1><m1| |m2><m2| ... |mn><mn| -> |m1>|m2>...|mn><m1|<m2|...<mn|
-    transpose_list = [2*i for i in range(size)] + [2*i + 1 for i in range(size)]
-    matview = np.transpose(mat, transpose_list).reshape((dim, dim))
-
-    # checks if matview is diagonal and, if so, use faster contractions
-    diag = np.all(matview == np.diag(np.diagonal(matview)))
-
-    if pure:
-        if n == 1:
-            if diag:
-                mat_diag = mat.diagonal()
-                return np.multiply(mat_diag, state)
-            return np.dot(mat, state)
-
-        # Transpose the state into the following form:
-        # |psi> |mode[0]> |mode[1]> ... |mode[n]>
-        transpose_list = [i for i in range(n) if not i in modes] + modes
-        view = np.transpose(state, transpose_list)
-
-        # Apply matrix to each substate
-        ret = np.zeros([trunc for i in range(n)], dtype=def_type)
-        for i in product(*([range(trunc) for j in range(n - size)])):
-            if diag:
-                mat_diag = matview.diagonal()
-                ret[i] = np.multiply(mat_diag, view[i].ravel()).reshape(stshape)
-            else:
-                ret[i] = np.dot(matview, view[i].ravel()).reshape(stshape)
-
-        # "untranspose" the return matrix ret
-        untranspose_list = [0] * len(transpose_list)
-        for i in range(len(transpose_list)): # pylint: disable=consider-using-enumerate
-            untranspose_list[transpose_list[i]] = i
-
-        return np.transpose(ret, untranspose_list)
-
-    # otherwise, the state is mixed
-    if n == 1:
-        if diag:
-            mat_diag = mat.diagonal().reshape(-1, 1)
-            return np.multiply(mat_diag, np.multiply(state, dagger(mat_diag)))
-        return np.dot(mat, np.dot(state, dagger(mat)))
-
-    # Transpose the state into the following form:
-    # |psi><psi||mode[0]>|mode[1]>...|mode[n]><mode[0]|<mode[1]|...<mode[n]|
-    transpose_list = [i for i in range(n*2) if not i//2 in modes]
-    transpose_list = transpose_list + [2*i for i in modes] + [2*i + 1 for i in modes]
-    view = np.transpose(state, transpose_list)
-
-    # Apply matrix to each substate
-    ret = np.zeros([trunc for i in range(n*2)], dtype=def_type)
-    for i in product(*([range(trunc) for j in range((n - size)*2)])):
-        if diag:
-            mat_diag = matview.diagonal().reshape(-1, 1)
-            ret[i] = np.multiply(mat_diag, np.multiply(view[i].reshape((dim, dim)), dagger(mat_diag))).reshape(stshape + stshape)
-        else:
-            ret[i] = np.dot(matview, np.dot(view[i].reshape((dim, dim)), dagger(matview))).reshape(stshape + stshape)
-
-    # "untranspose" the return matrix ret
-    untranspose_list = [0] * len(transpose_list)
-    for i in range(len(transpose_list)): # pylint: disable=consider-using-enumerate
-        untranspose_list[transpose_list[i]] = i
-
-    return np.transpose(ret, untranspose_list)
-
-
-def apply_gate_einsum(mat, state, pure, modes, n, trunc):
-    """
-    Gate application based on einsum.
-    Assumes the input matrix has shape (out1, in1, ...)
-    """
-    # pylint: disable=unused-argument
-
-    size = len(modes)
-
-    if pure:
-        if n == 1:
-            return np.dot(mat, state)
-
-        left_str = [indices[:size*2]]
-
-        j = genOfRange(size)
-        right_str = [indices[2*next(j) + 1] if i in modes else indices[size*2 + i] \
-            for i in range(n)]
-
-        j = genOfRange(size)
-        out_str = [indices[2*next(j)] if i in modes else indices[size*2 + i] \
-            for i in range(n)]
-
-        einstring = ''.join(left_str + [','] + right_str + ['->'] + out_str)
-        return np.einsum(einstring, mat, state)
-
-    # otherwise, the state is mixed
-    if n == 1:
-        return np.dot(mat, np.dot(state, dagger(mat)))
-
-    in_str = indices[:n*2]
-
-    j = genOfRange(n*2)
-    out_str = ''.join([indices[n*2 + next(j)] if i//2 in modes else indices[i] for i in range(n*2)])
-
-    j = genOfRange(size*2)
-    left_str = ''.join([out_str[modes[i//2]*2] if (i%2) == 0 else in_str[modes[i//2]*2] for i in range(size*2)])
-    right_str = ''.join([out_str[modes[i//2]*2 + 1] if (i%2) == 0 else in_str[modes[i//2]*2 + 1] for i in range(size*2)])
-
-    einstring = ''.join([left_str, ',', in_str, ',', right_str, '->', out_str])
-    return np.einsum(einstring, mat, state, mat.conj())
-
-
-def apply_twomode_gate(mat, state, pure, modes, n, trunc, gate="BSgate"):
-    """Applies a two-mode gate to a state.
-
-    Applies the specified two-mode gate to the state using custom tensor contractions and
-    the Numba compiler for faster application. Currently, only the beamsplitter and the
-    two-mode squeeze gate are supported.
-
-    Args:
-        mat (array[complex]): The numeric operator to be applied to the state, of shape `[trunc]*(2*n)`
-        state (array[complex]): The state that the operator is applied to
-        pure (bool): whether the state is pure or mixed
-        modes (list[int]): the list of modes to which the operator is applied on
-        n (int): The total number of modes
-        trunc (int): The Hilbert space truncation/cutoff
-        gate (str): The gate that is being applied. This argument determines the selection rules that
-            are used. Options are ``"BSgate"`` and ``"S2gate"``.
-
-    Returns:
-        array[complex]: state after application of the two-mode operation
-    """
-    if pure:
-        t1 = modes[0]
-        t2 = modes[1]
-
-        # moves the modes on which the gate should be applied to the
-        # front indices of the state, for mode 1 and 2 respectively
-        # e.g. from [i1, i2, i3] --> [i2, i3, i1] if the gate is applied to modes 2 and 3
-        switch_list_1 = np.arange(n)
-        switch_list_2 = np.arange(n)
-        switch_list_1[[0, t1]] = switch_list_1[[t1, 0]]
-        switch_list_2[[1, t2]] = switch_list_2[[t2, 1]]
-
-        state = state.transpose(switch_list_1)
-        state = state.transpose(switch_list_2)
-
-        if gate == "BSgate":
-            state = _apply_BS(mat, state, trunc)
-        elif gate == "S2gate":
-            state = _apply_S2(mat, state, trunc)
-        else:
-            raise NotImplementedError("Currently, selection rules are only implemented for the BSgate "
-                                      "and the S2gate. The {} gate is not supported".format(gate))
-
-        state = state.transpose(switch_list_2)
-        ret = state.transpose(switch_list_1)
-    else:
-        t1 = 2 * modes[0]
-        t2 = 2 * modes[1]
-
-        # moves the modes on which the gate should be applied to the
-        # front indices of the state, for mode 1 and 2 respectively
-        # e.g. from [i1, j1, i2, j2, i3, j3] --> [i2, j2, i3, j3, i1, j1]
-        # if the gate is applied to modes 2 and 3
-        switch_list_1 = np.arange(2 * n)
-        switch_list_2 = np.arange(2 * n)
-        switch_list_1[[0, 1, t1, t1+1]] = switch_list_1[[t1, t1+1, 0, 1]]
-        switch_list_2[[0, 1, t2, t2+1]] = switch_list_2[[t2, t2+1, 0, 1]]
-
-        # puts the modes on which the gate should be applied together
-        # e.g. from [i2, j2, i3, j3, i1, j1] --> [i2, i3, j2, j3, i1, j1]
-        # if the gate is applied to modes 2 and 3 (which are already switched to the first two modes)
-        transpose_list = np.arange(2 * n)
-        transpose_list[[t1+1, t2]] = transpose_list[[t2, t1+1]]
-
-        state = state.transpose(transpose_list)
-        state = state.transpose(switch_list_1)
-
-        if gate == "BSgate":
-            state = _apply_BS(mat, state, trunc)
-            state = state.transpose(switch_list_1)
-
-            state = state.transpose(switch_list_2)
-            state = _apply_BS(mat.conj(), state, trunc)
-
-        elif gate == "S2gate":
-            state = _apply_S2(mat, state, trunc)
-            state = state.transpose(switch_list_1)
-
-            state = state.transpose(switch_list_2)
-            state = _apply_S2(mat.conj(), state, trunc)
-
-        else:
-            raise NotImplementedError
-
-        state = state.transpose(switch_list_2)
-        ret = state.transpose(transpose_list)
-
-    return ret
-
-
-# ignored in covtest (doesn't work well with the jit decorator)
-@jit(nopython=True)
-def _apply_BS(mat, state, trunc):  # pragma: no cover
-    r"""Applies the BS gate to the first bra in state.
-
-    The beamsplitter matrix elements :math:`B_{ij}^{kl}` satisfy the selection
-    rules
-
-    .. math::
-        B_{ij}^{kl} \propto \delta_{i+j}^{k+l}
-
-    This implies that one can contract pure states (without loss of generality
-    we assume only two modes) using one less loop as follows:
-
-    .. math::
-        c'_{i, j} = \sum\limits_{k = \text{max}(1+i+j-n, 0)}
-                   ^{\text{min}(i+j, n-1) + 1}
-        B_{ik}^{j, i+j-k} c_{k, i+j-k}
-
-    where :math:`c_{kl}` is the tensor representing the two mode state and
-    :math:`n` is the cutoff in Fock space. Similarly for a mixed state it is
-    possible to do the update of the state using two less summations.
-    """
-    ret = np.zeros_like(state, dtype=np.complex128)
-    for i in range(trunc):
-        for j in range(trunc):
-            for k in range(max(1+i+j-trunc, 0), min(i+j, trunc-1) + 1):
-                ret[i, j] += mat[i, k, j, i+j-k] * state[k, i+j-k]
-    return ret
-
-
-# ignored in covtest (doesn't work well with the jit decorator)
-@jit(nopython=True)
-def _apply_S2(mat, state, trunc):  # pragma: no cover
-    r"""Applies the S2 gate to the first bra in state.
-
-    The two-mode squeeze matrix elements :math:`T_{ij}^{kl}` satisfy the selection
-    rules
-
-    .. math::
-        T_{ij}^{kl} \propto \delta_{i-j}^{k-l}
-
-    This implies that one can contract pure states (without loss of generality
-    we assume only two modes) using one less loop as follows:
-
-    .. math::
-        c'_{i, k} = \sum\limits_{k = \text{max}(i-j, 0)}^{\text{min}(i-j, 0) + n}
-        T_{ij}^{k, k+j-i} c_{j, k+j-i}
-
-    where :math:`c_{jl}` is the tensor representing the two mode state and
-    :math:`n` is the cutoff in Fock space. Similarly for a mixed state it is
-    possible to do the update of the state using two less summations.
-    """
-    ret = np.zeros_like(state, dtype=np.complex128)
-    for i in range(trunc):
-        for j in range(trunc):
-            for k in range(max(i-j, 0), trunc + min(i-j, 0)):
-                ret[i, k] += mat[i, j, k, k + j-i] * state[j, k + j-i]
     return ret
